refactor types into split files
