+++ packages/core/src/types/database.ts
@@ -1,0 +1,345 @@
+import type { Agent, Character } from './agent';
+import type { Component, Entity, Participant, Relationship, Room, World } from './environment';
+import type { Memory, MemoryMetadata } from './memory';
+import type { UUID } from './primitives';
+import type { Task } from './task';
+
+/**
+ * Represents a log entry
+ */
+export interface Log {
+  /** Optional unique identifier */
+  id?: UUID;
+
+  /** Associated entity ID */
+  entityId: UUID;
+
+  /** Associated room ID */
+  roomId?: UUID;
+
+  /** Log body */
+  body: { [key: string]: unknown };
+
+  /** Log type */
+  type: string;
+
+  /** Log creation timestamp */
+  createdAt: Date;
+}
+
+/**
+ * Interface for database operations
+ */
+export interface IDatabaseAdapter {
+  /** Database instance */
+  db: any;
+
+  /** Initialize database connection */
+  init(): Promise<void>;
+
+  /** Close database connection */
+  close(): Promise<void>;
+
+  getConnection(): Promise<any>;
+
+  getAgent(agentId: UUID): Promise<Agent | null>;
+
+  /** Get all agents */
+  getAgents(): Promise<Partial<Agent>[]>;
+
+  createAgent(agent: Partial<Agent>): Promise<boolean>;
+
+  updateAgent(agentId: UUID, agent: Partial<Agent>): Promise<boolean>;
+
+  deleteAgent(agentId: UUID): Promise<boolean>;
+
+  ensureEmbeddingDimension(dimension: number): Promise<void>;
+
+  /** Get entity by IDs */
+  getEntityByIds(entityIds: UUID[]): Promise<Entity[] | null>;
+
+  /** Get entities for room */
+  getEntitiesForRoom(roomId: UUID, includeComponents?: boolean): Promise<Entity[]>;
+
+  /** Create new entities */
+  createEntities(entities: Entity[]): Promise<boolean>;
+
+  /** Update entity */
+  updateEntity(entity: Entity): Promise<void>;
+
+  /** Get component by ID */
+  getComponent(
+    entityId: UUID,
+    type: string,
+    worldId?: UUID,
+    sourceEntityId?: UUID
+  ): Promise<Component | null>;
+
+  /** Get all components for an entity */
+  getComponents(entityId: UUID, worldId?: UUID, sourceEntityId?: UUID): Promise<Component[]>;
+
+  /** Create component */
+  createComponent(component: Component): Promise<boolean>;
+
+  /** Update component */
+  updateComponent(component: Component): Promise<void>;
+
+  /** Delete component */
+  deleteComponent(componentId: UUID): Promise<void>;
+
+  /** Get memories matching criteria */
+  getMemories(params: {
+    entityId?: UUID;
+    agentId?: UUID;
+    count?: number;
+    unique?: boolean;
+    tableName: string;
+    start?: number;
+    end?: number;
+    roomId?: UUID;
+    worldId?: UUID;
+  }): Promise<Memory[]>;
+
+  getMemoryById(id: UUID): Promise<Memory | null>;
+
+  getMemoriesByIds(ids: UUID[], tableName?: string): Promise<Memory[]>;
+
+  getMemoriesByRoomIds(params: {
+    tableName: string;
+    roomIds: UUID[];
+    limit?: number;
+  }): Promise<Memory[]>;
+
+  getCachedEmbeddings(params: {
+    query_table_name: string;
+    query_threshold: number;
+    query_input: string;
+    query_field_name: string;
+    query_field_sub_name: string;
+    query_match_count: number;
+  }): Promise<{ embedding: number[]; levenshtein_score: number }[]>;
+
+  log(params: {
+    body: { [key: string]: unknown };
+    entityId: UUID;
+    roomId: UUID;
+    type: string;
+  }): Promise<void>;
+
+  getLogs(params: {
+    entityId: UUID;
+    roomId?: UUID;
+    type?: string;
+    count?: number;
+    offset?: number;
+  }): Promise<Log[]>;
+
+  deleteLog(logId: UUID): Promise<void>;
+
+  searchMemories(params: {
+    embedding: number[];
+    match_threshold?: number;
+    count?: number;
+    unique?: boolean;
+    tableName: string;
+    query?: string;
+    roomId?: UUID;
+    worldId?: UUID;
+    entityId?: UUID;
+  }): Promise<Memory[]>;
+
+  createMemory(memory: Memory, tableName: string, unique?: boolean): Promise<UUID>;
+
+  updateMemory(memory: Partial<Memory> & { id: UUID; metadata?: MemoryMetadata }): Promise<boolean>;
+
+  deleteMemory(memoryId: UUID): Promise<void>;
+
+  deleteAllMemories(roomId: UUID, tableName: string): Promise<void>;
+
+  countMemories(roomId: UUID, unique?: boolean, tableName?: string): Promise<number>;
+
+  createWorld(world: World): Promise<UUID>;
+
+  getWorld(id: UUID): Promise<World | null>;
+
+  removeWorld(id: UUID): Promise<void>;
+
+  getAllWorlds(): Promise<World[]>;
+
+  updateWorld(world: World): Promise<void>;
+
+  getRoomsByIds(roomIds: UUID[]): Promise<Room[] | null>;
+
+  createRooms(rooms: Room[]): Promise<UUID[]>;
+
+  deleteRoom(roomId: UUID): Promise<void>;
+
+  deleteRoomsByWorldId(worldId: UUID): Promise<void>;
+
+  updateRoom(room: Room): Promise<void>;
+
+  getRoomsForParticipant(entityId: UUID): Promise<UUID[]>;
+
+  getRoomsForParticipants(userIds: UUID[]): Promise<UUID[]>;
+
+  getRoomsByWorld(worldId: UUID): Promise<Room[]>;
+
+  removeParticipant(entityId: UUID, roomId: UUID): Promise<boolean>;
+
+  getParticipantsForEntity(entityId: UUID): Promise<Participant[]>;
+
+  getParticipantsForRoom(roomId: UUID): Promise<UUID[]>;
+
+  addParticipantsRoom(entityIds: UUID[], roomId: UUID): Promise<boolean>;
+
+  getParticipantUserState(roomId: UUID, entityId: UUID): Promise<'FOLLOWED' | 'MUTED' | null>;
+
+  setParticipantUserState(
+    roomId: UUID,
+    entityId: UUID,
+    state: 'FOLLOWED' | 'MUTED' | null
+  ): Promise<void>;
+
+  /**
+   * Creates a new relationship between two entities.
+   * @param params Object containing the relationship details
+   * @returns Promise resolving to boolean indicating success
+   */
+  createRelationship(params: {
+    sourceEntityId: UUID;
+    targetEntityId: UUID;
+    tags?: string[];
+    metadata?: { [key: string]: any };
+  }): Promise<boolean>;
+
+  /**
+   * Updates an existing relationship between two entities.
+   * @param relationship The relationship object with updated data
+   * @returns Promise resolving to void
+   */
+  updateRelationship(relationship: Relationship): Promise<void>;
+
+  /**
+   * Retrieves a relationship between two entities if it exists.
+   * @param params Object containing the entity IDs and agent ID
+   * @returns Promise resolving to the Relationship object or null if not found
+   */
+  getRelationship(params: {
+    sourceEntityId: UUID;
+    targetEntityId: UUID;
+  }): Promise<Relationship | null>;
+
+  /**
+   * Retrieves all relationships for a specific entity.
+   * @param params Object containing the user ID, agent ID and optional tags to filter by
+   * @returns Promise resolving to an array of Relationship objects
+   */
+  getRelationships(params: { entityId: UUID; tags?: string[] }): Promise<Relationship[]>;
+
+  getCache<T>(key: string): Promise<T | undefined>;
+  setCache<T>(key: string, value: T): Promise<boolean>;
+  deleteCache(key: string): Promise<boolean>;
+
+  // Only task instance methods - definitions are in-memory
+  createTask(task: Task): Promise<UUID>;
+  getTasks(params: { roomId?: UUID; tags?: string[]; entityId?: UUID }): Promise<Task[]>;
+  getTask(id: UUID): Promise<Task | null>;
+  getTasksByName(name: string): Promise<Task[]>;
+  updateTask(id: UUID, task: Partial<Task>): Promise<void>;
+  deleteTask(id: UUID): Promise<void>;
+
+  getMemoriesByWorldId(params: {
+    worldId: UUID;
+    count?: number;
+    tableName?: string;
+  }): Promise<Memory[]>;
+}
+
+/**
+ * Result interface for embedding similarity searches
+ */
+export interface EmbeddingSearchResult {
+  embedding: number[];
+  levenshtein_score: number;
+}
+
+/**
+ * Options for memory retrieval operations
+ */
+export interface MemoryRetrievalOptions {
+  roomId: UUID;
+  count?: number;
+  unique?: boolean;
+  start?: number;
+  end?: number;
+  agentId?: UUID;
+}
+
+/**
+ * Options for memory search operations
+ */
+export interface MemorySearchOptions {
+  embedding: number[];
+  match_threshold?: number;
+  count?: number;
+  roomId: UUID;
+  agentId?: UUID;
+  unique?: boolean;
+  metadata?: Partial<MemoryMetadata>;
+}
+
+/**
+ * Options for multi-room memory retrieval
+ */
+export interface MultiRoomMemoryOptions {
+  roomIds: UUID[];
+  limit?: number;
+  agentId?: UUID;
+}
+
+/**
+ * Unified options pattern for memory operations
+ * Provides a simpler, more consistent interface
+ */
+export interface UnifiedMemoryOptions {
+  roomId: UUID;
+  limit?: number; // Unified naming (replacing 'count')
+  agentId?: UUID; // Common optional parameter
+  unique?: boolean; // Common flag for duplication control
+  start?: number; // Pagination start
+  end?: number; // Pagination end
+}
+
+/**
+ * Specialized memory search options
+ */
+export interface UnifiedSearchOptions extends UnifiedMemoryOptions {
+  embedding: number[];
+  similarity?: number; // Clearer name than 'match_threshold'
+}
+
+/**
+ * Represents a generic database connection object.
+ * The actual type of this connection will depend on the specific database adapter implementation
+ * (e.g., a connection pool object for PostgreSQL, a client instance for a NoSQL database).
+ * This `unknown` type serves as a placeholder in the abstract `IDatabaseAdapter`.
+ */
+export type DbConnection = unknown;
+
+/**
+ * A generic type for metadata objects, often used in various parts of the system like
+ * `Relationship` metadata or other extensible data structures.
+ * It allows for arbitrary key-value pairs where values are of `unknown` type,
+ * encouraging consumers to perform type checking or casting.
+ */
+export type MetadataObject = Record<string, unknown>;
+
+// Allowable vector dimensions
+export const VECTOR_DIMS = {
+  SMALL: 384,
+  MEDIUM: 512,
+  LARGE: 768,
+  XL: 1024,
+  XXL: 1536,
+  XXXL: 3072,
+} as const;
